chore(ws-core): lint & refactor
