+++ packages/graphql-server-ws/src/interfaces.ts
@@ -1,14 +1,14 @@
-import { GraphQLOptions } from "graphql-server-core";
+import { GraphQLOptions } from 'graphql-server-core';
 import { graphql, GraphQLSchema, ExecutionResult } from 'graphql';
 import { ReactiveExecuteFunction } from 'graphql-server-core';
 import * as WebSocket from 'ws';
 
 export interface WSRequest {
-	id: number; // Per socket increasing number
-	action: "request" | "cancel";
-	query?: string; // GraphQL Printed Query.
-	variables?: any; // GraphQL variables.
-	operationName?: string; // GraphQL operationName
+  id: number; // Per socket increasing number
+  action: 'request' | 'cancel';
+  query?: string; // GraphQL Printed Query.
+  variables?: any; // GraphQL variables.
+  operationName?: string; // GraphQL operationName
 }
 
 export interface WSMessageParams {
